fastapi/requirements.txt
fastapi==0.103.2
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pywatchman==1.4.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
